Decennial Estimates by Place

Generate decennial estimates for multiple tables by place(s).

get_decennial_place(table_codes, year, fips = NULL, state = "WA")

Arguments

table_codes

A character string or vector of Census table codes.

year

Numeric value. A decennial year equal or greater than 2010

fips

Character value. Single code or vector of place fips codes.

state

A character string state abbreviation

Value

a tibble of decennial estimates by place(s) for selected table codes. Does not include variable names.

Author

Christy Lam

Examples

Sys.getenv("CENSUS_API_KEY")
#> [1] "325779a5606507b511316d1ecd4328aa21cfc70d"
get_decennial_place(table_codes = 'PCT013', year = 2010)
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT013. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> Using Census Summary File 1
#> # A tibble: 30,772 × 4
#>    GEOID   NAME                         variable  value
#>    <chr>   <chr>                        <chr>     <dbl>
#>  1 5308780 Burbank CDP, Washington      PCT013001  3291
#>  2 5308850 Burien city, Washington      PCT013001 33013
#>  3 5308920 Burlington city, Washington  PCT013001  8280
#>  4 5309480 Camas city, Washington       PCT013001 19277
#>  5 5309810 Canterwood CDP, Washington   PCT013001  3079
#>  6 5310215 Carnation city, Washington   PCT013001  1786
#>  7 5309820 Canyon Creek CDP, Washington PCT013001  3200
#>  8 5309970 Carbonado town, Washington   PCT013001   610
#>  9 5310075 Carlsborg CDP, Washington    PCT013001   995
#> 10 5310320 Carson CDP, Washington       PCT013001  2279
#> # ℹ 30,762 more rows

get_decennial_place(table_codes = 'PCT013',
                    year = 2010,
                    fips = c("5363000", "5308850"))
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT013. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> Using Census Summary File 1
#> # A tibble: 98 × 4
#>    GEOID   NAME                     variable   value
#>    <chr>   <chr>                    <chr>      <dbl>
#>  1 5308850 Burien city, Washington  PCT013001  33013
#>  2 5363000 Seattle city, Washington PCT013001 583735
#>  3 5308850 Burien city, Washington  PCT013002  16612
#>  4 5363000 Seattle city, Washington PCT013002 290243
#>  5 5308850 Burien city, Washington  PCT013003   1140
#>  6 5363000 Seattle city, Washington PCT013003  16424
#>  7 5308850 Burien city, Washington  PCT013004   1065
#>  8 5363000 Seattle city, Washington PCT013004  12919
#>  9 5308850 Burien city, Washington  PCT013005    947
#> 10 5363000 Seattle city, Washington PCT013005  10989
#> # ℹ 88 more rows

get_decennial_place(table_codes = c('PCT013', 'PCT022'),
                    year = 2010,
                    fips = c("5363000", "5308850"))
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT013. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT022. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> # A tibble: 140 × 4
#>    GEOID   NAME                     variable   value
#>    <chr>   <chr>                    <chr>      <dbl>
#>  1 5308850 Burien city, Washington  PCT013001  33013
#>  2 5363000 Seattle city, Washington PCT013001 583735
#>  3 5308850 Burien city, Washington  PCT013002  16612
#>  4 5363000 Seattle city, Washington PCT013002 290243
#>  5 5308850 Burien city, Washington  PCT013003   1140
#>  6 5363000 Seattle city, Washington PCT013003  16424
#>  7 5308850 Burien city, Washington  PCT013004   1065
#>  8 5363000 Seattle city, Washington PCT013004  12919
#>  9 5308850 Burien city, Washington  PCT013005    947
#> 10 5363000 Seattle city, Washington PCT013005  10989
#> # ℹ 130 more rows